Created new UserAccountService class. Seperated other services. Added tests for roles, as well additional functionality (Error handling).

+++ identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/service/RoleService.java
@@ -1,31 +1,33 @@
 package nz.ac.canterbury.seng302.identityprovider.service;
 
-import io.grpc.stub.StreamObserver;
-import net.devh.boot.grpc.server.service.GrpcService;
 import nz.ac.canterbury.seng302.identityprovider.database.UserModel;
 import nz.ac.canterbury.seng302.identityprovider.database.UserRepository;
+import nz.ac.canterbury.seng302.identityprovider.exceptions.IrremovableRoleException;
+import nz.ac.canterbury.seng302.identityprovider.exceptions.UserDoesNotExistException;
 import nz.ac.canterbury.seng302.shared.identityprovider.ModifyRoleOfUserRequest;
-import nz.ac.canterbury.seng302.shared.identityprovider.UserAccountServiceGrpc;
 import nz.ac.canterbury.seng302.shared.identityprovider.UserRole;
 import nz.ac.canterbury.seng302.shared.identityprovider.UserRoleChangeResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 /**
  * This service handles adding and removing roles to/from users, as according to the
  * user_accounts.proto file.
  */
-@GrpcService
-public class RoleService extends UserAccountServiceGrpc.UserAccountServiceImplBase {
+@Service
+public class RoleService {
 
-  @Autowired
-  private UserRepository repository;
+  @Autowired private UserRepository repository;
 
   // Helper function to reduce duplicate code.
   // Gets the user, checks if modification is legal (returns false if not), modifies, saves.
   // If addingRole is true, role will be added. If false, role will be deleted.
-  private boolean modifyRoleOfUser(
-      ModifyRoleOfUserRequest modificationRequest, boolean addingRole) {
+  private boolean modifyRoleOfUser(ModifyRoleOfUserRequest modificationRequest, boolean addingRole)
+      throws UserDoesNotExistException {
     UserModel user = repository.findById(modificationRequest.getUserId());
+    if (user == null) {
+      throw new UserDoesNotExistException("Error: User does not exist.");
+    }
     UserRole roleToChange = modificationRequest.getRole();
 
     // If the role is being added, the user should not have the role.
@@ -34,6 +36,9 @@ public class RoleService extends UserAccountServiceGrpc.UserAccountServiceImplBa
       if (addingRole) {
         user.addRole(roleToChange);
       } else {
+        if (roleToChange == UserRole.STUDENT) {
+          throw new IrremovableRoleException("Error: Cannot remove STUDENT role.");
+        }
         user.deleteRole(roleToChange);
       }
       repository.save(user);
@@ -46,43 +51,33 @@ public class RoleService extends UserAccountServiceGrpc.UserAccountServiceImplBa
    * Adds a role to the user if the user does not already have the role.
    *
    * @param modificationRequest The modification request for the role
-   * @param responseObserver An observer to send back a response
+   * @return UserRoleChangeResponse The status of how successful the modification was
    */
-  @Override
-  public void addRoleToUser(
-      ModifyRoleOfUserRequest modificationRequest,
-      StreamObserver<UserRoleChangeResponse> responseObserver) {
+  public UserRoleChangeResponse addRoleToUser(ModifyRoleOfUserRequest modificationRequest)
+      throws UserDoesNotExistException {
     // Generate a response based on if modifyRoleOfUser was successful or not
     var response =
         modifyRoleOfUser(modificationRequest, true)
             ? UserRoleChangeResponse.newBuilder().setIsSuccess(true).setMessage(true)
             : UserRoleChangeResponse.newBuilder().setIsSuccess(false).setMessage(false);
-    // Return the response message
-    responseObserver.onNext(response.build());
-    responseObserver.onCompleted();
+
+    return response.build();
   }
 
   /**
    * Removes a role to the user if the user does not already have the role.
    *
    * @param modificationRequest The modification request for the role
-   * @param responseObserver An observer to send back a response
+   * @return UserRoleChangeResponse The status of how successful the modification was
    */
-  @Override
-  public void removeRoleFromUser(
-      ModifyRoleOfUserRequest modificationRequest,
-      StreamObserver<UserRoleChangeResponse> responseObserver) {
+  public UserRoleChangeResponse removeRoleFromUser(ModifyRoleOfUserRequest modificationRequest)
+      throws UserDoesNotExistException, IrremovableRoleException {
     // Generate a response based on if modifyRoleOfUser was successful or not
     var response =
         modifyRoleOfUser(modificationRequest, false)
             ? UserRoleChangeResponse.newBuilder().setIsSuccess(true).setMessage(true)
             : UserRoleChangeResponse.newBuilder().setIsSuccess(false).setMessage(false);
-    // Return the response message
-    responseObserver.onNext(response.build());
-    responseObserver.onCompleted();
-  }
 
-  public UserRepository getRepository() {
-    return repository;
+    return response.build();
   }
 }
